swe_env: collapse the spine modules 9 -> 5 (pure code-motion, no behavior change)
Fold small, single-purpose spine files into their cohesive owners so the library
is easier to navigate. No logic changes; only code relocation + import repoints.

- harness.py    <- absorbs registry.py + grading.py (the harness contract now also
                   holds the name->harness registry and the pure grading helpers).
- sandbox.py    <- environment.py renamed, absorbing lifecycle.py (the async sandbox
                   wrapper plus its acquire/teardown context manager).
- self_drive.py <- absorbs model_endpoint.py (self-driving scaffolding plus the
                   in-sandbox model-server egress primitive).
- verify_task.py / __init__.py / harnesses/ / tests/: import paths repointed.

Removed: registry.py, grading.py, lifecycle.py, model_endpoint.py (environment.py
renamed to sandbox.py). anyswe_agent/app.py unchanged.

ruff clean; swe_env + anyswe suites green (150 passed, 2 skipped).
