+++ responses_api_agents/swe_env/tests/test_lifecycle.py
@@ -28,7 +28,7 @@ import responses_api_agents.swe_env.harnesses  # noqa: F401  (register harnesses
 from nemo_gym.sandbox import SandboxExecResult, SandboxHandle, SandboxStatus
 from responses_api_agents.swe_env.harness import SweTask
 from responses_api_agents.swe_env.harnesses.swe_bench_ext import SweBenchExtHarness
-from responses_api_agents.swe_env.lifecycle import acquire_sandbox
+from responses_api_agents.swe_env.sandbox import acquire_sandbox
 from responses_api_agents.swe_env.verify_task import verify_task
